Clean up Run model files
