--- src/Client/Models/EnvironmentInfo.cs
@@ -1,8 +1,0 @@
-namespace Runly.Models
-{
-	public class EnvironmentInfo
-	{
-		public string Id { get; set; }
-		public bool IsProduction { get; set; }
-	}
-}
