--- src/Client/Models/NodeInfo.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Runly.Models
-{
-	public class NodeInfo
-	{
-		public Guid Id { get; set; }
-		public string Machine { get; set; }
-		public string Cluster { get; set; }
-	}
-}
